+++ java/vector/src/main/java/org/apache/arrow/vector/IntervalYearVector.java
@@ -117,7 +117,7 @@ public final class IntervalYearVector extends BaseFixedWidthVector {
    * @return value stored at the index.
    */
   public static int getTotalMonths(final ArrowBuf buffer, final int index) {
-    return buffer.getInt(index * TYPE_WIDTH);
+    return buffer.getInt((long) index * TYPE_WIDTH);
   }
 
   /**
@@ -130,7 +130,7 @@ public final class IntervalYearVector extends BaseFixedWidthVector {
     if (NULL_CHECKING_ENABLED && isSet(index) == 0) {
       throw new IllegalStateException("Value at index is null");
     }
-    return valueBuffer.getInt(index * TYPE_WIDTH);
+    return valueBuffer.getInt((long) index * TYPE_WIDTH);
   }
 
   /**
@@ -146,7 +146,7 @@ public final class IntervalYearVector extends BaseFixedWidthVector {
       return;
     }
     holder.isSet = 1;
-    holder.value = valueBuffer.getInt(index * TYPE_WIDTH);
+    holder.value = valueBuffer.getInt((long) index * TYPE_WIDTH);
   }
 
   /**
@@ -159,7 +159,7 @@ public final class IntervalYearVector extends BaseFixedWidthVector {
     if (isSet(index) == 0) {
       return null;
     } else {
-      final int interval = valueBuffer.getInt(index * TYPE_WIDTH);
+      final int interval = valueBuffer.getInt((long) index * TYPE_WIDTH);
       // TODO: verify interval is in months
       return Period.ofMonths(interval);
     }
@@ -181,7 +181,7 @@ public final class IntervalYearVector extends BaseFixedWidthVector {
   }
 
   private StringBuilder getAsStringBuilderHelper(int index) {
-    int value = valueBuffer.getInt(index * TYPE_WIDTH);
+    int value = valueBuffer.getInt((long) index * TYPE_WIDTH);
 
     final int years = (value / org.apache.arrow.vector.util.DateUtility.yearsToMonths);
     final int months = (value % org.apache.arrow.vector.util.DateUtility.yearsToMonths);
@@ -204,7 +204,7 @@ public final class IntervalYearVector extends BaseFixedWidthVector {
 
 
   private void setValue(int index, int value) {
-    valueBuffer.setInt(index * TYPE_WIDTH, value);
+    valueBuffer.setInt((long) index * TYPE_WIDTH, value);
   }
 
   /**
