+++ java/vector/src/main/java/org/apache/arrow/vector/UInt4Vector.java
@@ -83,7 +83,7 @@ public final class UInt4Vector extends BaseFixedWidthVector implements BaseIntVe
    * @return value stored at the index.
    */
   public static long getNoOverflow(final ArrowBuf buffer, final int index) {
-    long l = buffer.getInt(index * TYPE_WIDTH);
+    long l = buffer.getInt((long) index * TYPE_WIDTH);
     return (0x00000000FFFFFFFFL) & l;
   }
 
@@ -97,7 +97,7 @@ public final class UInt4Vector extends BaseFixedWidthVector implements BaseIntVe
     if (NULL_CHECKING_ENABLED && isSet(index) == 0) {
       throw new IllegalStateException("Value at index is null");
     }
-    return valueBuffer.getInt(index * TYPE_WIDTH);
+    return valueBuffer.getInt((long) index * TYPE_WIDTH);
   }
 
   /**
@@ -113,7 +113,7 @@ public final class UInt4Vector extends BaseFixedWidthVector implements BaseIntVe
       return;
     }
     holder.isSet = 1;
-    holder.value = valueBuffer.getInt(index * TYPE_WIDTH);
+    holder.value = valueBuffer.getInt((long) index * TYPE_WIDTH);
   }
 
   /**
@@ -126,7 +126,7 @@ public final class UInt4Vector extends BaseFixedWidthVector implements BaseIntVe
     if (isSet(index) == 0) {
       return null;
     } else {
-      return valueBuffer.getInt(index * TYPE_WIDTH);
+      return valueBuffer.getInt((long) index * TYPE_WIDTH);
     }
   }
 
@@ -153,7 +153,7 @@ public final class UInt4Vector extends BaseFixedWidthVector implements BaseIntVe
 
 
   private void setValue(int index, int value) {
-    valueBuffer.setInt(index * TYPE_WIDTH, value);
+    valueBuffer.setInt((long) index * TYPE_WIDTH, value);
   }
 
   /**
